requirements.txt
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
autopep8==2.0.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
cryptography==39.0.0
daphne==4.0.0
dj-database-url==1.2.0
Django==4.1.5
django-cors-headers==3.13.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httptools==0.5.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.2.4
msgpack==1.0.4
priority==1.3.0
prompt-toolkit==3.0.37
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tomli==2.0.1
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
whitenoise==6.3.0
zope.interface==5.5.2
